use milisecond precision

+++ internal/db/models.go
@@ -1,12 +1,12 @@
 package db
 
 type Detection struct {
-	ID                int64  `json:"id" gorm:"primaryKey,autoIncrement"`
-	DetectionTime     int64  `json:"detection_timestamp"`
-	Mac               string `json:"mac"`
-	Rssi              int    `json:"rssi"`
-	UptimeMs          uint64 `json:"uptime_ms"`
-	BatteryPercentage uint8  `json:"battery"`
+	ID                int64   `json:"id" gorm:"primaryKey,autoIncrement"`
+	DetectionTime     float64 `json:"detection_timestamp"`
+	Mac               string  `json:"mac"`
+	Rssi              int     `json:"rssi"`
+	UptimeMs          uint64  `json:"uptime_ms"`
+	BatteryPercentage uint8   `json:"battery"`
 }
 
 func (D *Detection) TableName() string {
